# server/openapi.yaml
openapi: 3.0.3
info:
  title: Geopolitical Dashboard API
  version: 1.1.0
  description: |
    Core backend API for datasets, portfolio analytics support data, reports, health checks,
    and news ingestion.
servers:
  - url: http://localhost:5001
paths:
...
  /api/workspace/state/{bucket}:
    get:
      summary: List latest persisted artifacts for a user/workspace bucket
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
            enum: [scenarioOutputs, alertConfigs, customThresholds, schedulesHistory, advancedPrefs]
        - in: query
          name: includeDeleted
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Latest versioned artifacts for bucket
        '503':
          description: Database unavailable
  /api/workspace/state/{bucket}/{artifactKey}:
    get:
      summary: Retrieve latest persisted artifact for bucket/key
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
            enum: [scenarioOutputs, alertConfigs, customThresholds, schedulesHistory, advancedPrefs]
        - in: path
          name: artifactKey
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest artifact
        '404':
          description: Artifact not found
    put:
      summary: Persist new artifact version with optimistic concurrency
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
            enum: [scenarioOutputs, alertConfigs, customThresholds, schedulesHistory, advancedPrefs]
        - in: path
          name: artifactKey
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Artifact version persisted
        '409':
          description: Version conflict
    delete:
      summary: Soft-delete artifact by creating a tombstone version
      parameters:
        - in: path
          name: bucket
          required: true
          schema:
            type: string
            enum: [scenarioOutputs, alertConfigs, customThresholds, schedulesHistory, advancedPrefs]
        - in: path
          name: artifactKey
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact deleted
        '404':
          description: Artifact not found
  /api/reports/generate:
    post:
      summary: Generate report metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Generation accepted
        '429':
          description: Expensive endpoint abuse protection triggered
  /api/reports/email:
    post:
      summary: Queue report email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Email queued
        '429':
          description: Expensive endpoint abuse protection triggered
...
  /api/reports/schedule:
    post:
      summary: Schedule recurring report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Schedule created
        '429':
          description: Expensive endpoint abuse protection triggered
...